refactor wave item selection router
