+++ src/app/router_registry.py
@@ -317,6 +317,7 @@ from app.routers.dpm_wave_campaign_workflow_boards import (
 )
 from app.routers.dpm_wave_detail import router as dpm_wave_detail_router
 from app.routers.dpm_wave_evidence import router as dpm_wave_evidence_router
+from app.routers.dpm_wave_item_selection import router as dpm_wave_item_selection_router
 from app.routers.dpm_wave_items import router as dpm_wave_items_router
 from app.routers.dpm_wave_lifecycle_actions import router as dpm_wave_lifecycle_actions_router
 from app.routers.dpm_wave_lookup import router as dpm_wave_lookup_router
@@ -642,6 +643,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_wave_lookup_router)
     app.include_router(dpm_wave_detail_router)
     app.include_router(dpm_wave_items_router)
+    app.include_router(dpm_wave_item_selection_router)
     app.include_router(dpm_wave_actions_router)
     app.include_router(dpm_wave_lifecycle_actions_router)
     app.include_router(dpm_wave_workflow_actions_router)
